# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST multi-session para WhatsApp usando whatsmeow. Baseada na WuzAPI com campos JSON em PascalCase.",
        "title": "FioZap API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "FioZap Support",
            "url": "https://github.com/fiozap/fiozap"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/sessions/{name}/profile/business/resolve": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Resolve um link de mensagem comercial (nao implementado)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Resolver link comercial",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Link",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResolveBusinessLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/profile/qrlink": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna link QR para adicionar contato (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Obter QR link de contato",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/profile/qrlink/resolve": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Resolve um link QR de contato (nao implementado)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Resolver QR link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Link QR",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResolveContactQRRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/profile/status": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Altera o recado/status do perfil (nao implementado)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Alterar recado",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novo status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SetStatusMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/sessions/{name}/users/{phone}/business": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna perfil comercial de um contato (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Perfil comercial",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Numero do telefone",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ResolveBusinessLinkRequest": {
            "type": "object",
            "properties": {
                "Link": {
                    "type": "string"
                }
            }
        },
        "dto.ResolveContactQRRequest": {
            "type": "object",
            "properties": {
                "Link": {
                    "type": "string"
                }
            }
        },
        "dto.Response": {
            "description": "Resposta padrao da API FioZap",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "string",
                    "example": "error message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.SetStatusMessageRequest": {
            "type": "object",
            "properties": {
                "Status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token global (GLOBAL_API_TOKEN) ou token da sessao",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}